refactor: merge OptionalAuthProvider and AuthProvider
