+++ client/src/App.tsx
@@ -3,7 +3,6 @@ import HelloWeb3 from "./components/pages/HelloWeb3";
 import SignUp from "./components/pages/SignUp";
 import SignIn from "./components/pages/SignIn";
 import Mypage from "./components/pages/Mypage";
-import OptionalAuthProvider from "./components/organisms/providers/OptionalAuthProvider";
 import AuthProvider from "./components/organisms/providers/AuthProvider";
 import ImageUpload from "./components/pages/ImageUpload";
 import Header from "./components/organisms/Header";
@@ -12,18 +11,27 @@ import NotFound from "./components/pages/NofFound";
 const App = () => {
   return (
     <>
-      <OptionalAuthProvider>
-        <Header />
-      </OptionalAuthProvider>
+      <Header />
       <hr />
       <Routes>
         <Route path="/" element={<HelloWeb3 />} />
         <Route path="/signup" element={<SignUp />} />
         <Route path="/signin" element={<SignIn />} />
-        <Route path="/mypage" element={<AuthProvider element={<Mypage />} />} />
+        <Route
+          path="/mypage"
+          element={
+            <AuthProvider>
+              <Mypage />
+            </AuthProvider>
+          }
+        />
         <Route
           path="/image-upload"
-          element={<AuthProvider element={<ImageUpload />} />}
+          element={
+            <AuthProvider>
+              <ImageUpload />
+            </AuthProvider>
+          }
         />
         <Route path="*" element={<NotFound />} />
       </Routes>
