+++ client/src/components/organisms/Header/index.tsx
@@ -1,11 +1,19 @@
-import { useOptionalAuth } from "~/components/organisms/providers/OptionalAuthProvider";
+import AuthProvider, {
+  useAuthContext,
+} from "~/components/organisms/providers/AuthProvider";
+
+const AcountInfo = () => {
+  const { user } = useAuthContext();
+  return <p>{user.getUsername()}</p>;
+};
 
 const Header = () => {
-  const auth = useOptionalAuth();
   return (
     <div>
       <h1>knowtfolio</h1>
-      <p>{auth && auth.user.getUsername()}</p>
+      <AuthProvider contentForUnauthorized={<p>未認証ユーザー</p>}>
+        <AcountInfo />
+      </AuthProvider>
     </div>
   );
 };
